+++ demo/server.yml
@@ -1,15 +1,16 @@
-ui: true
+ui: true                               # UI enabled
 
-iface: "wg0"
-network: "192.168.2.0/24"
+iface: "wg0"                           # Wireguard interface to use
 
 client:
-    enabled: false
+    enabled: false                     # Client enabled
 
 server:
-    enabled: true
-    bind_addr: "127.0.0.1"
-    secret: "seashell-test-secret-123"
+    enabled: true                      # Server enabled
+    bind_addr: "127.0.0.1"             # Server bind address
+    network: "192.168.2.0/24"          # Overlay network address space
+    secret: "seashell-test-secret-123" # Secret used to sign JWT tokens issued to clients
+
