--- wg0.conf
@@ -1,17 +1,0 @@
-[Interface]
-Address = 192.168.2.1/24
-ListenPort = 51820
-PrivateKey = UOqqP7mFGaG6qY77bwFyHiTU5RukzaUvZgXO1RP9hUI=
-
-[Peer]
-PublicKey = i2tdRKD7ObUrfLmJ4y6ZuBFvLLSTJe1mQvCZELYaqng=
-AllowedIPs = 192.168.2.3/24
-PersistentKeepalive = 20
-
-[Peer]
-PublicKey = nAjpVPq6nS8SpPw7MbajIjIckTun6CwNnSZD30M9njo=
-AllowedIPs = 192.168.2.2/24
-PersistentKeepalive = 20
-
-
-
